Add a new Main Page Component. Move all the initial app component to this new component. Set initial app path to the main page component.
