+++ src/app/app-routing.module.ts
@@ -1,9 +1,11 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { AboutPageComponent } from './pages/about-page/about-page.component';
+import { MainPageComponent } from './pages/main-page/main-page.component';
 
 
 const routes: Routes = [
+  { path: '', component: MainPageComponent },
   { path: 'about', component: AboutPageComponent },
 ];
